Tentativa de passar dados da API por props
