+++ pokedex/src/pages/PokedexPage/PokedexPage.js
@@ -8,20 +8,7 @@ import { useState } from 'react';
 
 
 function PokedexPage(props) {
-  const [pokedexList, setPokedexList] = useState(props.pokedex)
-
-  const addPokemon = (pokemon) => {
-    if (!pokedexList.includes(pokemon)) {
-      const updatedList = [...pokedexList, pokemon];
-      setPokedexList(updatedList);
-    }
-  }
-
-  const removePokemon = (pokemon) => {
-    const pokemonIndex = pokedexList.indexOf(pokemon);
-    const updatedList = pokedexList.splice(pokemonIndex, 1);
-    setPokedexList(updatedList);
-  }
+  const [pokedexList, setPokedexList] = useState(props.pokedexList)
 
   const pokemonsCards = pokedexList && pokedexList.map(
     pokemon => {
@@ -31,13 +18,11 @@ function PokedexPage(props) {
           cardWidth={250}
           textBtn1={"remover"}
           textBtn2={"detalhes"}
-          onClickBtn1={() => removePokemon(pokemon.name)}
-          onClickBtn2={() => addPokemon(pokemon.name)}
           colorBtn1={"primary"}
           colorBtn2={"secondary"}
           variantBtn1={"contained"}
           variantBtn2={"contained"}
-
+          onClickBtn1={() => props.removePokemon(pokemon.name)}
         />
       )
     }
@@ -47,8 +32,8 @@ function PokedexPage(props) {
     <div>
       <Box sx={{ display: 'grid', rowGap: 4, gridTemplateColumns: 'repeat(4, 2fr)', m: 4, p: 2 }}>
         {pokemonsCards}
-        <button onClick={() => props.addPokemon(pokemon)}>Adicionar Pokemon</button>
       </Box>
+      {/* PAGINAÇÃO */}
       <Box display="flex" justifyContent="center" sx={{ m: 5 }}>
         <Stack spacing={2}>
           <Pagination count={10} color="secondary" />
